backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
drf-spectacular==0.27.2
django-cors-headers==4.6.0
psycopg[binary]==3.2.3
gunicorn==23.0.0
python-dotenv==1.0.1
